+++ Sources/KeyboardKitSwiftUI/Extensions/Color+ClearInteractable.swift
@@ -13,7 +13,7 @@ public extension Color {
      This color can be used instead of `.clear`, which makes
      a view stop registering touches and gestures.
      */
-    static var clearInteractable: some View {
-        Color.white.opacity(0.001)
+    static var clearInteractable: Color {
+        Color(white: 0.001)
     }
 }