+++ Tests/KeyboardKitSwiftUITests/System/View+SystemTests.swift
@@ -1,5 +1,5 @@
 //
-//  View+KeyboardStylesTests.swift
+//  View+System.swift
 //  KeyboardKitTests
 //
 //  Created by Daniel Saidi on 2020-06-24.
@@ -12,7 +12,7 @@ import KeyboardKit
 import KeyboardKitSwiftUI
 import SwiftUI
 
-class View_KeyboardStylesTests: QuickSpec {
+class View_SystemKeyboardStylesTests: QuickSpec {
 
     override func spec() {
         
@@ -50,24 +50,8 @@ class View_KeyboardStylesTests: QuickSpec {
         describe("system keyboard button background") {
             
             it("is defined") {
-                let background = Text("").systemKeyboardButtonBackground(for: .dark, appearance: .light)
+                let background = Text("").systemKeyboardButtonBackground(for: .backspace, scheme: .dark, appearance: .light)
                 expect(background).toNot(beNil())
-            }
-        }
-        
-        describe("system keyboard button background color") {
-            
-            it("varies for different schemes and appearances") {
-                let lightLight = Text("").systemKeyboardButtonBackgroundColor(for: .light, appearance: .light)
-                let lightDark = Text("").systemKeyboardButtonBackgroundColor(for: .light, appearance: .dark)
-                let darkLight = Text("").systemKeyboardButtonBackgroundColor(for: .dark, appearance: .light)
-                let darkDark = Text("").systemKeyboardButtonBackgroundColor(for: .dark, appearance: .dark)
-                expect(lightLight).toNot(equal(lightDark))
-                expect(lightLight).toNot(equal(darkLight))
-                expect(lightLight).toNot(equal(darkDark))
-                expect(lightDark).toNot(equal(darkLight))
-                expect(lightDark).toNot(equal(darkDark))
-                expect(darkLight).to(equal(darkDark))
             }
         }
         
@@ -103,20 +87,6 @@ class View_KeyboardStylesTests: QuickSpec {
             }
         }
         
-        describe("system keyboard button foreground color") {
-            
-            it("varies for different schemes and appearances") {
-                let lightLight = Text("").systemKeyboardButtonForegroundColor(for: .light, appearance: .light)
-                let lightDark = Text("").systemKeyboardButtonForegroundColor(for: .light, appearance: .dark)
-                let darkLight = Text("").systemKeyboardButtonForegroundColor(for: .dark, appearance: .light)
-                let darkDark = Text("").systemKeyboardButtonForegroundColor(for: .dark, appearance: .dark)
-                expect(lightLight).toNot(equal(lightDark))
-                expect(lightLight).to(equal(darkLight))
-                expect(lightLight).to(equal(darkDark))
-                expect(lightDark).to(equal(.white))
-            }
-        }
-        
         describe("system keyboard button frame") {
             
             it("is defined") {
@@ -137,30 +107,6 @@ class View_KeyboardStylesTests: QuickSpec {
             
             it("is defined") {
                 let value = Text("").systemKeyboardButtonShadow(for: .dark, appearance: .light)
-                expect(value).toNot(beNil())
-            }
-        }
-        
-        describe("system keyboard button shadow color") {
-            
-            it("is defined") {
-                let value = Text("").systemKeyboardButtonShadowColor
-                expect(value).toNot(beNil())
-            }
-        }
-        
-        describe("system keyboard button text shadow") {
-            
-            it("is defined") {
-                let value = Text("").systemKeyboardButtonTextShadow(for: .dark, appearance: .default)
-                expect(value).toNot(beNil())
-            }
-        }
-        
-        describe("system keyboard button text shadow color") {
-            
-            it("is defined") {
-                let value = Text("").systemKeyboardButtonTextShadowColor
                 expect(value).toNot(beNil())
             }
         }